# Conflicts:
#   src/main/java/com/example/fams/repository/LearningObjectiveContentRepository.java
#   src/main/java/com/example/fams/repository/SyllabusMaterialRepository.java

+++ src/main/java/com/example/fams/services/impl/ClassServiceImpl.java
@@ -1,14 +1,21 @@
 package com.example.fams.services.impl;
 
+import com.example.fams.config.CustomValidationException;
 import com.example.fams.config.ResponseUtil;
 import com.example.fams.converter.GenericConverter;
 import com.example.fams.dto.ClassDTO;
+import com.example.fams.dto.ContentDTO;
+import com.example.fams.dto.LearningObjectiveDTO;
 import com.example.fams.dto.UserDTO;
+import com.example.fams.entities.Content;
 import com.example.fams.entities.FamsClass;
 import com.example.fams.entities.ClassUser;
+import com.example.fams.entities.LearningObjective;
+import com.example.fams.entities.LearningObjectiveContent;
 import com.example.fams.entities.User;
 import com.example.fams.repository.ClassRepository;
 import com.example.fams.repository.ClassUserRepository;
+import com.example.fams.repository.UserClassRepository;
 import com.example.fams.repository.UserRepository;
 import com.example.fams.services.IClassService;
 import com.example.fams.services.ServiceUtils;
@@ -25,13 +32,15 @@ import java.util.List;
 public class ClassServiceImpl implements IClassService {
 
   private final ClassRepository classRepository;
+  private final UserClassRepository userClassRepository;
   private final ClassUserRepository classUserRepository;
   private final UserRepository userRepository;
   private final GenericConverter genericConverter;
 
-  public ClassServiceImpl(ClassRepository classRepository, ClassUserRepository classUserRepository,
-      UserRepository userRepository, GenericConverter genericConverter) {
+  public ClassServiceImpl(ClassRepository classRepository, UserClassRepository userClassRepository,
+      ClassUserRepository classUserRepository, UserRepository userRepository, GenericConverter genericConverter) {
     this.classRepository = classRepository;
+    this.userClassRepository = userClassRepository;
     this.classUserRepository = classUserRepository;
     this.userRepository = userRepository;
     this.genericConverter = genericConverter;
@@ -44,24 +53,7 @@ public class ClassServiceImpl implements IClassService {
     List<FamsClass> entities = classRepository.findAllByStatusIsTrue(pageable);
     List<ClassDTO> result = new ArrayList<>();
 
-
-    for (FamsClass entity : entities) {
-      ClassDTO newClassDTO = (ClassDTO) genericConverter.toDTO(entity, ClassDTO.class);
-
-
-      List<User> users = classUserRepository.findUserByClassId(entity.getId());
-
-      List<UserDTO> userDTOs = new ArrayList<>();
-      for (User user : users) {
-        UserDTO newUserDTO = (UserDTO) genericConverter.toDTO(user, UserDTO.class);
-        userDTOs.add(newUserDTO);
-      }
-
-      newClassDTO.setUserDTOs(userDTOs);
-
-      result.add(newClassDTO);
-    }
-
+    convertListClassToListClassDTO(entities, result);
 
     return ResponseUtil.getCollection(result,
         HttpStatus.OK,
@@ -69,22 +61,31 @@ public class ClassServiceImpl implements IClassService {
         page,
         limit,
         classRepository.countAllByStatusIsTrue());
+
   }
 
   @Override
   public ResponseEntity<?> save(ClassDTO classDTO) {
-    List<UserDTO> requestUserDTOs = classDTO.getUserDTOs();
-
+    ServiceUtils.errors.clear();
+    List<Long> requestUserIds = classDTO.getUserIds();
     FamsClass entity;
+
+    // * Validate requestDTO ( if left null, then can be updated later )
+    if (requestUserIds != null){
+      ServiceUtils.validateUserIds(requestUserIds, userRepository);
+    }
+    if (!ServiceUtils.errors.isEmpty()) {
+      throw new CustomValidationException(ServiceUtils.errors);
+    }
 
     // * For update request
     if (classDTO.getId() != null){
       FamsClass oldEntity = classRepository.findById(classDTO.getId());
       FamsClass tempOldEntity = ServiceUtils.cloneFromEntity(oldEntity);
-      entity = (FamsClass) genericConverter.updateEntity(classDTO, oldEntity);
-      entity = ServiceUtils.fillMissingAttribute(entity, tempOldEntity);
+      entity = convertDtoToEntity(classDTO);
+      ServiceUtils.fillMissingAttribute(entity, tempOldEntity);
       classUserRepository.deleteAllByFamsClassId(classDTO.getId());
-      loadClassUserFromListUserId(requestUserDTOs, entity.getId());
+      loadClassUserFromListUserId(requestUserIds, entity.getId());
       entity.markModified();
       classRepository.save(entity);
     }
@@ -94,28 +95,21 @@ public class ClassServiceImpl implements IClassService {
       classDTO.setStatus(true);
       entity = (FamsClass) genericConverter.toEntity(classDTO, FamsClass.class);
       classRepository.save(entity);
-      loadClassUserFromListUserId(requestUserDTOs, entity.getId());
+      loadClassUserFromListUserId(requestUserIds, entity.getId());
     }
 
 
     ClassDTO result = (ClassDTO) genericConverter.toDTO(entity, ClassDTO.class);
-
-    List<User> users = classUserRepository.findUserByClassId(entity.getId());
-    List<UserDTO> userDTOS = new ArrayList<>();
-    for (User user : users) {
-      UserDTO newUserDTO = (UserDTO) genericConverter.toDTO(user, UserDTO.class);
-      userDTOS.add(newUserDTO);
-    }
-    result.setUserDTOs(userDTOS);
+    result.setUserIds(requestUserIds);
     return ResponseUtil.getObject(result, HttpStatus.OK, "Saved successfully");
   }
 
-  private void loadClassUserFromListUserId(List<UserDTO> requestUserDTOs, Long classId) {
-    if (requestUserDTOs != null && !requestUserDTOs.isEmpty()) {
-      for (UserDTO userDTO : requestUserDTOs) {
+  private void loadClassUserFromListUserId(List<Long> requestUserIds, Long classId) {
+    if (requestUserIds != null && !requestUserIds.isEmpty()) {
+      for (Long userId : requestUserIds) {
         ClassUser clu = new ClassUser();
         clu.setFamsClass(classRepository.findById(classId));
-        clu.setUser(userRepository.findById(userDTO.getId()).get());
+        clu.setUser(userRepository.findById(userId).get());
         classUserRepository.save(clu);
       }
     }
@@ -136,20 +130,11 @@ public class ClassServiceImpl implements IClassService {
   @Override
   public ResponseEntity<?> findAll(int page, int limit) {
     Pageable pageable = PageRequest.of(page - 1, limit);
-    List<FamsClass> entities = classRepository.findAllByOrderByIdDesc(pageable);
+    List<FamsClass> entities = classRepository.findAllBy(pageable);
     List<ClassDTO> result = new ArrayList<>();
-    for (FamsClass entity : entities) {
-      ClassDTO newDTO = (ClassDTO) genericConverter.toDTO(entity, ClassDTO.class);
-      List<User> users = classUserRepository.findUserByClassId(entity.getId());
-      List<UserDTO> userDTOs = new ArrayList<>();
-      for(User us : users){
-        UserDTO userDTO = (UserDTO) genericConverter.toDTO(us,UserDTO.class);
-        userDTOs.add(userDTO);
-      }
-      newDTO.setUserDTOs(userDTOs);
 
-      result.add(newDTO);
-    }
+    convertListClassToListClassDTO(entities, result);
+
     return ResponseUtil.getCollection(result,
         HttpStatus.OK,
         "Fetched successfully",
@@ -178,17 +163,11 @@ public class ClassServiceImpl implements IClassService {
   public ResponseEntity<?> searchSortFilter(ClassDTO classDTO, int page, int limit) {
     String code = classDTO.getCode();
     String name = classDTO.getName();
-//    Long duration = classDTO.getDuration();
-//    Long startDate = classDTO.getStartDate();
-//    Long endDate = classDTO.getEndDate();
     Pageable pageable = PageRequest.of(page - 1, limit);
-    List<FamsClass> entities = classRepository.searchSortFilter(code, name, /*duration, startDate, endDate,*/ pageable);
+    List<FamsClass> entities = classRepository.searchSortFilter(code, name, pageable);
     List<ClassDTO> result = new ArrayList<>();
-    Long count = classRepository.countSearchSortFilter(code, name/*, duration, startDate, endDate*/);
-    for (FamsClass entity : entities){
-      ClassDTO newDTO = (ClassDTO) genericConverter.toDTO(entity, ClassDTO.class);
-      result.add(newDTO);
-    }
+    Long count = classRepository.countSearchSortFilter(code, name);
+    convertListClassToListClassDTO(entities, result);
     return ResponseUtil.getCollection(result,
         HttpStatus.OK,
         "Fetched successfully",
@@ -205,13 +184,10 @@ public class ClassServiceImpl implements IClassService {
 //    Long startDate = classDTO.getStartDate();
 //    Long endDate = classDTO.getEndDate();
     Pageable pageable = PageRequest.of(page - 1, limit);
-    List<FamsClass> entities = classRepository.searchSortFilterADMIN(code, name/*, duration, startDate, endDate*/, sortById, pageable);
+    List<FamsClass> entities = classRepository.searchSortFilterADMIN(code, name, sortById,  pageable);
     List<ClassDTO> result = new ArrayList<>();
-    Long count = classRepository.countSearchSortFilter(code, name/*, duration, startDate, endDate*/);
-    for (FamsClass entity : entities){
-      ClassDTO newDTO = (ClassDTO) genericConverter.toDTO(entity, ClassDTO.class);
-      result.add(newDTO);
-    }
+    Long count = classRepository.countSearchSortFilter(code, name);
+    convertListClassToListClassDTO(entities, result);
     return ResponseUtil.getCollection(result,
         HttpStatus.OK,
         "Fetched successfully",
@@ -219,4 +195,37 @@ public class ClassServiceImpl implements IClassService {
         limit,
         count);
   }
+
+  private void convertListClassToListClassDTO(List<FamsClass> entities, List<ClassDTO> result) {
+    for (FamsClass famsClass : entities){
+      ClassDTO newClassDTO = (ClassDTO) genericConverter.toDTO(famsClass, ClassDTO.class);
+      List<User> users = classUserRepository.findUserByClassId(famsClass.getId());
+      if (famsClass.getClassUsers() == null){
+        newClassDTO.setUserIds(null);
+      }
+      else {
+        // ! Set list learningObjectiveIds và unitId sau khi convert ở trên vào contentDTO
+        List<Long> userIds = users.stream()
+            .map(User::getId)
+            .toList();
+
+
+        newClassDTO.setUserIds(userIds);
+
+      }
+      result.add(newClassDTO);
+    }
+  }
+
+  public FamsClass convertDtoToEntity(ClassDTO classDTO) {
+    FamsClass famsClass = new FamsClass();
+    famsClass.setId(classDTO.getId());
+    famsClass.setName(classDTO.getName());
+    famsClass.setCode(classDTO.getCode());
+    famsClass.setStatus(classDTO.getStatus());
+
+    return famsClass;
+  }
+
+
 }
